feat: 변수명 수정
